Foundry/master test ux (#295)
* Update ci.yml

Signed-off-by: Dargon789 <64915515+Dargon789@users.noreply.github.com>

* Update ci.yml (#211)


This pull request updates the Rust version in the CircleCI workflow to 1.89.0. This is a good maintenance task to keep the CI environment up-to-date. I have one suggestion regarding the Docker image tag to potentially simplify future maintenance by automatically adopting patch releases. Overall, the change is correct and beneficial.

Signed-off-by: Dargon789 <64915515+Dargon789@users.noreply.github.com>

* Update test.yml (#250)

CI:
Include the 'main' branch in the push event triggers for the test workflow
Signed-off-by: Dargon789 <64915515+Dargon789@users.noreply.github.com>

---------

Signed-off-by: Dargon789 <64915515+Dargon789@users.noreply.github.com>